Expand flask_app test coverage

+++ mltsp/Flask/flask_app.py
@@ -311,12 +311,6 @@ def check_user_table():
                        .count().run(g.rdb_conn) == 0:
 
         add_user()
-
-        print("User", stormpath.user.full_name, "with email",
-              stormpath.user.email, "added to users db.")
-    else:
-        print("User", stormpath.user.full_name, "with email",
-              stormpath.user.email, "already in users db.")
     return jsonify({})
 
 
@@ -1553,7 +1547,7 @@ def update_project_info(
     for fpath in delete_fpaths:
         try:
             os.remove(fpath)
-        except:
+        except Exception as e:
             pass
     new_proj_details = get_project_details(new_name)
     return new_proj_details
@@ -2017,7 +2011,8 @@ def prediction_proc(
     for i in range(n_cols_html_table):
         results_str += (
             "            <th class='pred_results'>Class%d</th>"
-            "            <th class='pred_results'>Class%d_Prob</th>") % (i + 1, i + 1)
+            "            <th class='pred_results'>Class%d_Prob</th>") % (i + 1,
+                                                                         i + 1)
     results_str += (
         "        </tr>"
         "    </thead>"
@@ -2139,6 +2134,8 @@ def verifyNewScript():
 
     """
     if request.method == "POST":
+        if 'custom_feat_script_file' not in request.files:
+            return "No custom features script uploaded. Please try again."
         scriptfile = request.files['custom_feat_script_file']
         scriptfile_name = secure_filename(scriptfile.filename)
         scriptfile_path = os.path.join(
@@ -2163,7 +2160,9 @@ def verifyNewScript():
             print(theErr)
             logging.exception("verifyNewScript error.")
             return str(theErr)
-        os.remove(scriptfile_path)
+        finally:
+            if os.path.exists(scriptfile_path):
+                os.remove(scriptfile_path)
         return str(
             "The following features have successfully been tested: <br>"
             + res_str)
@@ -2266,7 +2265,7 @@ def newProject(
             })
         proj_description = str(request.form["project_description"]).strip()
         addl_users = str(request.form["addl_authed_users"]).strip().split(',')
-        if addl_users == [''] or addl_users == [' ']:
+        if addl_users in [[''], ["None"]]:
             addl_users = []
         if "user_email" in request.form:
             user_email = str(request.form["user_email"]).strip()
@@ -2313,8 +2312,8 @@ def editOrDeleteProject():
             # return Response(status=str(result))
         else:
             print ("###### ERROR - editOrDeleteProject() - 'action' not "
-                   "in ['Edit','Delete'] ########")
-            return Response()
+                   "in ['Edit', 'Delete'] ########")
+            return jsonify({"error": "Invalid request action."})
 
 
 @app.route(
